+++ src/main/java/com/technovaperu/technovaperuwebstore/model/dto/create/CrearLoteDTO.java
@@ -40,6 +40,6 @@ public class CrearLoteDTO {
     @NotNull(message = "El cantidad no puede ser nula")
     @Positive(message = "El cantidad debe ser mayor que 0")
     @Schema(description = "Cantidad del lote", example = "100")
-    int cantidad;
+    BigDecimal cantidad;
     
 }
