+++ src/main/java/com/technovaperu/technovaperuwebstore/model/dto/update/ActualizarLoteDTO.java
@@ -41,6 +41,6 @@ public class ActualizarLoteDTO {
     @NotNull(message = "El cantidad no puede ser nula")
     @Positive(message = "El cantidad debe ser mayor que 0")
     @Schema(description = "Cantidad del lote", example = "100")
-    int cantidad;
+    BigDecimal cantidad;
     
 }
